moves API call for userinfo to home container and adds programList feature from local redux state; GET programsByUserID currently not working
