+++ src/components/Program/ProgramList.js
@@ -10,12 +10,10 @@ const ProgramList = () => {
     <div>
       <div>
         <h2>My Programs</h2>
-        <Link to="/add-program">Add Program</Link>
-        Add Program
       </div>
       <div>
         {programs.map(program => {
-          return <ProgramCard key={program.id} info={program} />;
+          return <ProgramCard key={program.id} programToEdit={program} />;
         })}
       </div>
     </div>
